refacto: removed everything useless in mmu.rs in order to put it in their own structs
