+++ src/cpu.rs
@@ -16,16 +16,16 @@ use std::fmt;
 use std::rc::Rc;
 
 use crate::cpu::registers::{R8, R16, Registers};
-use crate::memory::MemoryBus;
+use crate::mmu::Mmu;
 
 pub struct Cpu {
     pub registers: Registers,
     pub pc: u16,
-    pub bus: Rc<RefCell<MemoryBus>>,
+    pub bus: Rc<RefCell<Mmu>>,
 }
 
 impl Cpu {
-    pub fn new(bus: Rc<RefCell<MemoryBus>>) -> Self {
+    pub fn new(bus: Rc<RefCell<Mmu>>) -> Self {
         Cpu {
             registers: Registers::default(),
             bus,
@@ -107,7 +107,7 @@ impl Default for Cpu {
     fn default() -> Self {
         Cpu {
             registers: Registers::default(),
-            bus: Rc::new(RefCell::new(MemoryBus::default())),
+            bus: Rc::new(RefCell::new(Mmu::default())),
             pc: 0x0100,
         }
     }
@@ -123,7 +123,7 @@ mod tests {
 
     fn run_rom_test(rom_path: &str, logfile_name: &str) {
         let rom_data = fs::read(rom_path).expect("Failed to read ROM file");
-        let bus = Rc::new(RefCell::new(MemoryBus::new(rom_data)));
+        let bus = Rc::new(RefCell::new(Mmu::new(rom_data)));
         let mut cpu = Cpu::new(bus.clone());
         let mut logfile = fs::File::create(format!("logfiles/{}", logfile_name))
             .expect("Failed to create logfile");
