+++ src/gameboy.rs
@@ -5,19 +5,19 @@ use std::cell::RefCell;
 use std::rc::Rc;
 
 use crate::cpu::Cpu;
-use crate::memory::MemoryBus;
+use crate::mmu::Mmu;
 use crate::ppu::Ppu;
 
 #[derive(Default)]
 pub struct GameBoy {
     pub cpu: Cpu,
     pub ppu: Ppu,
-    pub bus: Rc<RefCell<MemoryBus>>,
+    pub bus: Rc<RefCell<Mmu>>,
 }
 
 impl GameBoy {
     pub fn new(rom: Vec<u8>) -> Self {
-        let bus = Rc::new(RefCell::new(MemoryBus::new(rom)));
+        let bus = Rc::new(RefCell::new(Mmu::new(rom)));
         let cpu = Cpu::new(bus.clone());
         let ppu = Ppu::new(bus.clone());
