--- src/memory/mod.rs
@@ -1,4 +1,0 @@
-pub mod bus;
-pub mod region;
-
-pub use bus::MemoryBus;
